HomeWork6/OwnIterator/src/main/java/org/example/ArrayIterator.java
package org.example;

import java.util.Iterator;

public class ArrayIterator<T> implements Iterator<T> {

    private final T[] array;
    private int currentIndex;

    public ArrayIterator(T[] array) {
        this.array = array;
        this.currentIndex = 0;
    }

    @Override
    public boolean hasNext() {
        return currentIndex < array.length;
    }

    @Override
    public T next() {
        if (!hasNext()) {
            throw new java.util.NoSuchElementException();
        }
        return array[currentIndex++];
    }

    public int getCurrentIndex() {
        return currentIndex;
    }
}
